Mirror of EA WebKit 16.4.1.0.2 from https://gpl.ea.com/eawebkit.html

+++ WebCore/platform/network/ea/TransportHandlerDirtySDKEA.h
@@ -74,8 +74,8 @@ namespace TransportHelper
 }
 }
 
-#include "protohttpmanager.h"
-#include "dirtyvers.h" // Defines (e.g.) #define DIRTYVERS (0x07000000)
+#include <DirtySDK/proto/protohttpmanager.h>
+#include <DirtySDK/dirtyvers.h> // Defines (e.g.) #define DIRTYVERS (0x07000000)
 
 
 namespace EA
@@ -102,11 +102,11 @@ public:
 	bool Tick			();
 
 #if DIRTYVERS > 0x07050300 
-	static int32_t DirtySDKSendHeaderCallbackStatic(ProtoHttpRefT* pState, char* pHeader, uint32_t uHeaderSize, const char* pData, uint32_t uDataLen, void* pUserRef);
+    static int32_t DirtySDKSendHeaderCallbackStatic(ProtoHttpRefT* pState, char* pHeader, uint32_t uHeaderSize, const char* pData, int64_t uDataLen, void* pUserRef);
 #else
 	static void DirtySDKSendHeaderCallbackStatic(ProtoHttpRefT* pState, char* pHeader, uint32_t uHeaderSize, const char* pData, uint32_t uDataLen, void* pUserRef);
 #endif
-	int32_t        DirtySDKSendHeaderCallback(ProtoHttpRefT* pState, char* pHeader, uint32_t uHeaderSize, const char* pData, uint32_t uDataLen, EA::WebKit::TransportInfo* pTInfo);
+    int32_t        DirtySDKSendHeaderCallback(ProtoHttpRefT* pState, char* pHeader, uint32_t uHeaderSize, const char* pData, uint32_t uDataLen, EA::WebKit::TransportInfo* pTInfo);
 
 #if DIRTYVERS > 0x07050300
 	static int32_t DirtySDKRecvHeaderCallbackStatic(ProtoHttpRefT* pState, const char* pHeader, uint32_t uHeaderSize, void* pUserRef);
